% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFastCache.R
\docType{class}
\name{RFastCache}
\alias{RFastCache}
\title{RFastCache

Cache class for R with set, get, has, delete and clear functions.}
\format{
An \link{'R6Class`} generator object.
}
\usage{
mycache <- RFastCache$new('/path/to/cache')
... FIXME
}
\description{
An R6 class that chaches an R object.
}
\examples{
mycache = RFastCache$new(dir="tmp/mycache")
bigdatatable=data.frame(x=c(1,2,3),y=c(1,4,9))
mycache$set('mykey', bigdatatable, 10)
print(mycache$get('long', default="long not cached anymore"))

for(i in 1:20) {
  print(mycache$get('mykey', default="mykey not cached anymore"))
  Sys.sleep(1); #1s
}

mycache$set('long', bigdatatable, 10*60)

## ------------------------------------------------
## Method `RFastCache$new`
## ------------------------------------------------

cache = RFastCache$new(my/new/cache)

## ------------------------------------------------
## Method `RFastCache$set`
## ------------------------------------------------

My_value = c(1:100)
cache$set(key = "My_Key", value = My_value, 1000)

## ------------------------------------------------
## Method `RFastCache$get`
## ------------------------------------------------

value = cache$get(key = "My_Key", default = "Sorry, but no value")

## ------------------------------------------------
## Method `RFastCache$has`
## ------------------------------------------------

is_value_there = cache$has(key = "My_Key")

## ------------------------------------------------
## Method `RFastCache$delete`
## ------------------------------------------------

cache$delete(key = "My_Key")

## ------------------------------------------------
## Method `RFastCache$clear`
## ------------------------------------------------

cache$clear()
}
\keyword{cache,}
\keyword{disk}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RFastCache$new()}}
\item \href{#method-set}{\code{RFastCache$set()}}
\item \href{#method-get}{\code{RFastCache$get()}}
\item \href{#method-has}{\code{RFastCache$has()}}
\item \href{#method-delete}{\code{RFastCache$delete()}}
\item \href{#method-clear}{\code{RFastCache$clear()}}
\item \href{#method-clone}{\code{RFastCache$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initializes R6 cache object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RFastCache$new(dir)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dir}}{Directory for cache.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A R6 RFastCache object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{cache = RFastCache$new(my/new/cache)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set"></a>}}
\if{latex}{\out{\hypertarget{method-set}{}}}
\subsection{Method \code{set()}}{
Method for adding objects to the cache
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RFastCache$set(key, value, ttl = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{By this key object will be stored in cache}

\item{\code{value}}{The value to be stored}

\item{\code{ttl}}{Time to live until stored object is outdated}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{My_value = c(1:100)
cache$set(key = "My_Key", value = My_value, 1000)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get"></a>}}
\if{latex}{\out{\hypertarget{method-get}{}}}
\subsection{Method \code{get()}}{
Method for geting objects from the cache
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RFastCache$get(key, default = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{The key by which the cached value will be retrieved}

\item{\code{default}}{If no value with the given key exist, set the value to default}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns value by key, if no value exist returns default
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{value = cache$get(key = "My_Key", default = "Sorry, but no value")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-has"></a>}}
\if{latex}{\out{\hypertarget{method-has}{}}}
\subsection{Method \code{has()}}{
Method for looking up objects from the cache
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RFastCache$has(key)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{The key by which the cached value will be looked up}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
true if value exist, false if no value or if value is outdated
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{is_value_there = cache$has(key = "My_Key")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete"></a>}}
\if{latex}{\out{\hypertarget{method-delete}{}}}
\subsection{Method \code{delete()}}{
Method for removing objects from the cache by key
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RFastCache$delete(key)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{The key of value}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{cache$delete(key = "My_Key")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clear"></a>}}
\if{latex}{\out{\hypertarget{method-clear}{}}}
\subsection{Method \code{clear()}}{
Method for clearing outdated objects from the cache
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RFastCache$clear()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{cache$clear()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RFastCache$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
